ai-productivity-suite/requirements.txt
fastapi
uvicorn
imapclient
caldav
stripe
appwrite
qdrant-client
langchain
langchain-community
autogen
python-dotenv